[SSHOC-467] Spring upgrade to 3.5.8
